+++ InworldAI/Config/DefaultInworldAI.ini
@@ -9,3 +9,4 @@
 +StructRedirects=(OldName="/Script/InworldAIIntegration.FInworldAuth",NewName="/Script/InworldAIClient.FInworldAuth")
 +StructRedirects=(OldName="/Script/InworldAIIntegration.FInworldSessionToken",NewName="/Script/InworldAIClient.FInworldSessionToken")
 +StructRedirects=(OldName="/Script/InworldAIIntegration.FInworldEnvironment",NewName="/Script/InworldAIClient.FInworldEnvironment")
++ClassRedirects=(OldName="/Script/InworldAIClient.UInworldBlueprintFunctionLibrary",NewName="/Script/InworldAIIntegration.UInworldBlueprintFunctionLibrary")
